requirements.txt
flask
flask-assets
blinker
markdown
markdown-checklist
markdown-include
markdown-full-yaml-metadata